Move example profiles to separate file
