+++ src/plugins/discover/public/context_awareness/profiles/index.ts
@@ -9,3 +9,5 @@
 export * from './root_profile';
 export * from './data_source_profile';
 export * from './document_profile';
+
+import('./example_profiles');
